Move zscore_nifti into utils.utils.
